rename old admin_screen to election_manager_screen, both mark and mark-scan
