+++ apps/mark-scan/frontend/src/pages/election_manager_screen.tsx
