+++ apps/mark/frontend/src/app_root.tsx
@@ -60,7 +60,7 @@ import {
 import { Ballot } from './components/ballot';
 import * as GLOBALS from './config/globals';
 import { BallotContext } from './contexts/ballot_context';
-import { AdminScreen } from './pages/admin_screen';
+import { AdminScreen } from './pages/election_manager_screen';
 import { InsertCardScreen } from './pages/insert_card_screen';
 import { PollWorkerScreen } from './pages/poll_worker_screen';
 import { SetupPrinterPage } from './pages/setup_printer_page';
