+++ apps/mark/frontend/src/pages/election_manager_screen.test.tsx
@@ -15,7 +15,7 @@ import { screen, within } from '../../test/react_testing_library';
 import { render } from '../../test/test_utils';
 import { election, defaultPrecinctId } from '../../test/helpers/election';
 
-import { AdminScreen, AdminScreenProps } from './admin_screen';
+import { AdminScreen, AdminScreenProps } from './election_manager_screen';
 import { mockMachineConfig } from '../../test/helpers/mock_machine_config';
 import {
   ApiMock,
