+++ apps/mark/frontend/src/pages/election_manager_screen.tsx
